(int) in spec val added

+++ RPG_autoBattler/Effects/Burning.cs
@@ -10,7 +10,7 @@ namespace RPG_autoBattler
 
         public void Trigger(TriggerType triggerType, Character attacker, Character victim, float[] specVal)
         {
-            if ((triggerType == TriggerType.EndTurn) && (specVal[0] == 0) && (TurnsLeft > 0))
+            if ((triggerType == TriggerType.EndTurn) && ((int)specVal[0] == 0) && (TurnsLeft > 0))
             {
                 Console.WriteLine($"{attacker} is on fire! {TurnsLeft - 1} turns left!");
                 TurnsLeft--;
